# docs/swagger.yaml
basePath: /
definitions:
  api.BatchSetConfigReq:
    properties:
      items:
        additionalProperties:
          type: string
        type: object
    required:
    - items
    type: object
...
  api.ChatReq:
    properties:
      botId:
        type: string
      text:
        type: string
    required:
    - botId
    - text
    type: object
...
  api.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  api.SetConfigReq:
    properties:
      value:
        type: string
    required:
    - value
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: 多渠道 AI 聊天机器人框架的 HTTP API
  title: thinkbot API
  version: "1.0"
paths:
...
  /api/chat/bots:
    get:
      description: 返回当前处于运行状态的 Bot 列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - CookieAuth: []
      summary: 可聊 Bot 列表
      tags:
      - 聊天
...
  /api/chat/send:
    post:
      consumes:
      - application/json
      description: 向指定 Bot 发送消息，通过 SSE 流式返回回复
      parameters:
      - description: 聊天请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.ChatReq'
      produces:
      - text/event-stream
      responses:
        "200":
          description: SSE 事件流
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - CookieAuth: []
      summary: 发送消息（SSE）
      tags:
      - 聊天
  /api/config:
    get:
      description: 列出所有系统配置项（需要 system.config 权限）
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - CookieAuth: []
      summary: 获取所有配置
      tags:
      - 系统配置
    put:
      consumes:
      - application/json
      description: 批量设置多个系统配置项
      parameters:
      - description: 批量配置请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.BatchSetConfigReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - CookieAuth: []
      summary: 批量设置配置
      tags:
      - 系统配置
  /api/config/{key}:
    put:
      consumes:
      - application/json
      description: 设置单个系统配置项的值
      parameters:
      - description: 配置键
        in: path
        name: key
        required: true
        type: string
      - description: 配置值请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.SetConfigReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - CookieAuth: []
      summary: 设置配置项
      tags:
      - 系统配置
...
securityDefinitions:
  CookieAuth:
    in: cookie
    name: token
    type: apiKey
swagger: "2.0"